server/pyproject.toml
[tool.poetry]
name = "qradviser"
version = "0.0.1"
description = "QRCode Bus Stop or Station adviser "
authors = ["Maruo.S <maruo@marusoftware.net>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = "^0.115.4"
tortoise-orm = "^0.21.6"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.5.2"
itsdangerous = "^2.2.0"
aerich = "^0.7.2"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.12"
gtfs-realtime-bindings = "^1.0.0"

[tool.poetry.group.dev.dependencies]
fastapi = {extras = ["dev"], version = "^0.115.4"}
watchfiles = "^0.24.0"
tortoise-cli = "^0.1.2"


[tool.poetry.group.postgres.dependencies]
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.6"}



[tool.poetry.group.getdata.dependencies]
python-odpt = "^0.1.3"

[tool.aerich]
tortoise_orm = "qradviser.db.DB_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
